:truck: Rename `data` package
