+++ app/src/main/java/network/omisego/omgmerchant/pages/authorized/confirm/handler/AbstractQRHandler.kt
@@ -2,7 +2,7 @@ package network.omisego.omgmerchant.pages.authorized.confirm.handler
 
 import android.arch.lifecycle.MutableLiveData
 import co.omisego.omisego.model.APIError
-import network.omisego.omgmerchant.data.RemoteRepository
+import network.omisego.omgmerchant.repository.RemoteRepository
 import network.omisego.omgmerchant.livedata.Event
 import network.omisego.omgmerchant.model.APIResult
 import network.omisego.omgmerchant.model.Feedback
