+++ app/src/main/java/network/omisego/omgmerchant/pages/authorized/main/more/account/SettingAccountViewModel.kt
@@ -12,8 +12,8 @@ import android.arch.lifecycle.ViewModel
 import co.omisego.omisego.model.Account
 import co.omisego.omisego.model.params.AccountWalletListParams
 import network.omisego.omgmerchant.base.StateViewHolderBinding
-import network.omisego.omgmerchant.data.LocalRepository
-import network.omisego.omgmerchant.data.RemoteRepository
+import network.omisego.omgmerchant.repository.LocalRepository
+import network.omisego.omgmerchant.repository.RemoteRepository
 import network.omisego.omgmerchant.databinding.ViewholderSettingAccountBinding
 import network.omisego.omgmerchant.extensions.mutableLiveDataOf
 import network.omisego.omgmerchant.livedata.Event
