+++ app/src/main/java/network/omisego/omgmerchant/repository/RemoteRepository.kt
@@ -1,4 +1,4 @@
-package network.omisego.omgmerchant.data
+package network.omisego.omgmerchant.repository
 
 /*
  * OmiseGO
